prevent delete if catewgory or subcategory have products

+++ laravel-api/app/Http/Controllers/Api/CategoryController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Api;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\Category\CreateCategoryRequest;
 use App\Http\Resources\CategoryResource;
+use App\Http\Resources\ProductCategoryResource;
 use App\Models\Category;
 use App\Models\Product;
 use Illuminate\Http\Request;
@@ -23,8 +24,30 @@ class CategoryController extends Controller
      */
     public function index(): \Illuminate\Http\JsonResponse
     {
-        $productCategory = Category::whereNull('parent_id')->get();
-        return (CategoryResource::collection($productCategory))->response();
+        /*$productCategory = cache()->remember('categories', 3600, function () {
+            return
+        });*/
+        $productCategory = Category::whereNull('parent_id')
+            ->withCount('products')
+            ->with(['children.children' => function ($query) {
+                $query->withCount('products');
+            }])
+            //}, 'children.products'])
+            //->with('products')
+            ->get();
+
+        /*foreach ($productCategory as $parentCategory) {
+            foreach ($parentCategory->children as $category) {
+                $category->products_count = $category->children->sum('products_count');
+            }
+            $parentCategory->products_count = $parentCategory->children->sum('products_count');
+        }*/
+
+        //return response()->json(['data' => $productCategory], 200);
+        return (ProductCategoryResource::collection($productCategory))->response();
+
+        /*return (CategoryResource::collection($productCategory))->response();
+        $productCategory = Category::whereNull('parent_id')->get();*/
     }
 
     /**
@@ -90,25 +113,48 @@ class CategoryController extends Controller
      * Remove the specified resource from storage.
      *
      * @param int $id
-     * @return CategoryResource
+     * @return ProductCategoryResource
      */
     public function destroy($id)
     {
-        $category = Category::findOrFail($id);
+        //$category = Category::findOrFail($id);
 
-        $proCounter = $category->products->count() ?: 0;
-        if (count($category->children)) {
-            foreach ($category->children as $child) {
-                if ($child->products->count()) {
-                    $proCounter += $child->products->count();
-                }
-            }
+        $category = Category::where('id', $id)
+            ->withCount('products')
+            ->with(['children.children' => function ($query) {
+                $query->withCount('products');
+            }])
+            ->first();
+
+        // Check ... Products count for ... selected category
+        if ($category->products_count > 0) {
+            return response()->json(["errorMsg" => "Cannot delete category! There are products linked to this category!"], 500);
         }
+
+        // Check ... Products count for ... SubCategories of selected category
+        $proCounter = 0;
+        $proCounter = $this->loopCategories($category, $proCounter);
+        //return response()->json(["errorMsg" => $proCounter, 'ukggb' => $category->children], 200);
 
         if ($proCounter > 0) {
             return response()->json(["errorMsg" => "Cannot delete category! There are products linked to this category or it's subcategories!"], 500);
         }
         abort_unless($category && $category->forceDelete(), 500, 'Unable to delete category!');
-        return new CategoryResource($category);
+        return new ProductCategoryResource($category);
+    }
+
+    public function loopCategories($category, $proCounter)
+    {
+        if (count($category->children) > 0) {
+            foreach ($category->children as $child) {
+                if ($child->products->count()) {
+                    $proCounter += $child->products->count();
+                }
+                if (count($child->children) > 0) {
+                    $proCounter += $this->loopCategories($child, $proCounter);
+                }
+            }
+        }
+        return $proCounter;
     }
 }
